+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to the project will be documented here.
 The format is based on [Keep a Changelog v1.1.0](https://keepachangelog.com/en/1.1.0/),
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added a check for `requires_sudo` to `run_all_tasks()` and `main()`.
+
 ## [0.6.1] - 2025-10-22
 
 ### Added